Huge Update
Updated the bootstrapper, added a camera to the characters, added steam networking support, added in animations for the beggining, fixed and cleaned up the character controller so network and client side are handled seperately. Added a scene management system for both client side and server side and also made it possible to load singleplayer levels with no server or client server stuff needed.